backend/requirements.txt
asgiref==3.8.1
beautifulsoup4==4.12.3
certifi==2024.7.4
charset-normalizer==3.3.2
Django==5.1
django-cors-headers==4.4.0
django-rest-framework==0.1.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
fix-busted-json==0.0.18
idna==3.7
psycopg2-binary==2.9.9
PyJWT==2.9.0
requests==2.32.3
soupsieve==2.6
sqlparse==0.5.1
urllib3==2.2.2
